Loading data from the server works now.
This change turned out to be much bigger than I anticipated.
To do this correctly, I need to recalculate the display data
each time render is called. This could be expensive, so I
started looking at handling things in getDerivedStateFromProps
but the react community seems to think this is the wrong approach
for things. I basically wanted to do memoization, so thats exactly
what I implemented. This explains the memoize-one dependency.
I added lodash because I didn't want to reinvent the isEqual
implementation which I need for memoization.

Additional Info:

* https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#what-about-memoization